Rename alias folder to alias_free_activation
