+++ alias_free_activation/cuda/compat.h
